+++ cpp/subprojects/common/include/mlrl/common/input/feature_matrix_csr.hpp
@@ -27,16 +27,17 @@ class MLRLCOMMON_API ICsrFeatureMatrix : public IRowWiseFeatureMatrix {
  *                      values
  * @param indices       A pointer to an array of type `uint32`, shape `(numNonZeroValues)`, that stores the
  *                      column-indices, the values in `values` correspond to
- * @param indptr        A pointer to an array of type `uint32`, shape `(numRows + 1)`, that stores the indices of the
- *                      first element in `values` and `indices` that corresponds to a certain row. The index at the last
+ * @param indptr        A pointer to an array of type `uint32`, shape `(numRows + 1)`, that stores the indices of first
+ *                      element in `values` and `indices` that corresponds to a certain row. The index at the last
  *                      position is equal to `numNonZeroValues`
  * @param numRows       The number of rows in the feature matrix
  * @param numCols       The number of columns in the feature matrix
+ * @param sparseValue   The value that should be used for sparse elements in the feature matrix
  * @return              An unique pointer to an object of type `ICsrFeatureMatrix` that has been created
  */
 MLRLCOMMON_API std::unique_ptr<ICsrFeatureMatrix> createCsrFeatureMatrix(const float32* values, uint32* indices,
-                                                                         uint32* indptr, uint32 numRows,
-                                                                         uint32 numCols);
+                                                                         uint32* indptr, uint32 numRows, uint32 numCols,
+                                                                         float32 sparseValue = 0.0f);
 
 #ifdef _WIN32
     #pragma warning(pop)